auth

+++ routes/api.php
@@ -4,6 +4,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\PostController;
 use App\Http\Resources\User as UserResource;
+use Illuminate\Support\Facades\Auth;
 
 /*
 |--------------------------------------------------------------------------
@@ -21,12 +22,15 @@ use App\Http\Resources\User as UserResource;
 Route::middleware('auth:sanctum')->get('/user', function (Request $request){
     return new UserResource($request->user());
 });
+Route::middleware('auth:sanctum')->get('/testauth', function (Request $request){
+    return Auth::user();
+});
 
 //Post
 Route::get('/posts/categories', [PostController::class, 'categories']);
 Route::get('/posts', [PostController::class, 'index']);
 Route::get('/posts/{id}', [PostController::class, 'show']);
 Route::get('/posts/category/{name}', [PostController::class, 'category']);
-Route::post('/posts/store', [PostController::class, 'store']);  //->middleware('auth:sanctum')
-Route::put('/posts/{id}', [PostController::class, 'update']);
-Route::delete('/posts/{id}', [PostController::class, 'delete']);
+Route::post('/posts/store', [PostController::class, 'store'])->middleware('auth:sanctum');  //->middleware('auth:sanctum')
+Route::put('/posts/{id}', [PostController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('/posts/{id}', [PostController::class, 'delete'])->middleware('auth:sanctum');
